JavaFX Client/Server Program
JavaFX application that deploys a client and server UI. The client connects to the Server over a socket and checks to see if a number is prime. If the user accidentally enters a letter/String, an error dialog will open. An executable JAR file was generated to deploy from the command line.