JavaFX Client/Server Prime Game
The Client asks the user if a number is prime over a socket connection. The server replies with the answer shown on its console. The user inputs a string text instead of an integer, an alert will display. The executable JAR will deploy both Client and Server applications. I used initialization methods from within the main.application to make this all work. Enjoy!